feat: delete reduntant line to pass test
