+++ controllers/reply-controller.js
@@ -1,21 +1,35 @@
 const { User, Reply, Tweet } = require('../models')
-const tweetServices = require('../services/tweet-services')
+const helpers = require('../_helpers')
 const handlebars = require('handlebars')
 handlebars.registerHelper('dateFormat', require('handlebars-dateformat'))
 
 const replyController = {
   getReply: (req, res, next) => {
-    return Tweet.findByPk(req.params.tweetId, {
-      include: [
-        { model: Reply, include: User },
-        { model: User }
-      ]
-    })
-      .then(tweet => {
-        console.log(req.params.tweetId)
-        console.log(tweet)
+    Promise.all([
+      Tweet.findByPk(req.params.id, {
+        include: [
+          { model: Reply, include: User },
+          { model: User }
+        ]
+      }),
+      User.findAll({
+        include: [{ model: User, as: 'Followers' }]
+      })
+    ])
+      .then(([tweet, users]) => {
+      // .then(tweet => {
         if (!tweet) throw new Error("tweet didn't exist!")
-        tweetServices.getFollowing(req, (err, users) => err ? next(err) : res.render('reply', { users: users.users, tweet: tweet.toJSON() }))
+        users = users.map(user => ({
+          ...user.toJSON()
+        }))
+        const filterSelfAndAdminUser = []
+        users = users.forEach(user => {
+          if (user.id !== helpers.getUser(req).id && user.role !== 'admin') {
+            filterSelfAndAdminUser.push(user)
+          }
+        })
+        users = filterSelfAndAdminUser.sort((a, b) => b.followerCount - a.followerCount).slice(0, 10)
+        res.render('reply', { tweet: tweet.toJSON() })
       })
       .catch(err => next(err))
   }
