+++ routes/modules/tweets.js
@@ -4,7 +4,7 @@ const router = express.Router()
 const replyController = require('../../controllers/reply-controller')
 const tweetController = require('../../controllers/tweet-controller')
 
-router.get('/:tweetId/replies', replyController.getReply)
+router.get('/:id/replies', replyController.getReply)
 router.get('/', tweetController.getTweets)
 
 module.exports = router
