+++ views/index.hbs
@@ -7,7 +7,7 @@
       <img src="images/img.png" alt="Can't 顯示" style="width: 50px;" class="me-3">
       <div>
         <p>Apple <span class="text-muted">@apple・3 小時</span></p>
-        <a href="/tweets/2/replies" style="text-decoration: none; color:black;">
+        <a href="/tweets/1/replies" style="text-decoration: none; color:black;">
           <p>Nulla Lorem mollit cupidatat irure. Laborum magna nulla duis ullamco cillum dolor.
             Voluptate exercitation incididunt aliquip deserunt reprehenderit elit laborum.<small>This line of text is
               meant to be treated as fine print.</small></p>
